Reorganise tests by behaviour
