+++ tests/Unit/DataTransferObjectTest.php
@@ -7,7 +7,6 @@ namespace Rexlabs\DataTransferObject\Tests\Unit;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use Rexlabs\DataTransferObject\DataTransferObject;
-use Rexlabs\DataTransferObject\Exceptions\UnknownPropertiesError;
 use Rexlabs\DataTransferObject\Property;
 
 use function spl_object_id;
@@ -48,7 +47,7 @@ class DataTransferObjectTest extends TestCase
      * @test
      * @return void
      */
-    public function uses_value_if_set(): void
+    public function access_property_if_defined(): void
     {
         $object = new DataTransferObject(
             [
@@ -61,158 +60,6 @@ class DataTransferObjectTest extends TestCase
         self::assertEquals('value', $object->__get('one'));
     }
 
-    /**
-     * @test
-     * @return void
-     */
-    public function get_unknown_property_throws(): void
-    {
-        $this->expectException(UnknownPropertiesError::class);
-
-        $object = new DataTransferObject([], [], NONE);
-
-        $object->__get('blim');
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function set_unknown_property_throws(): void
-    {
-        $this->expectException(UnknownPropertiesError::class);
-
-        $object = new DataTransferObject([], [], MUTABLE);
-
-        $object->__set('blim', 'blam');
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_property_reverts_to_default(): void
-    {
-        $type = $this->createMock(Property::class);
-        $type->method('processDefault')->willReturn('blam');
-
-        $object = new DataTransferObject(
-            ['blim' => $type],
-            [],
-            NONE
-        );
-
-        self::assertEquals('blam', $object->__get('blim'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function defined_property_returns_isset_true(): void
-    {
-        $object = new DataTransferObject(
-            ['blim' => $this->createMock(Property::class)],
-            ['blim' => true],
-            NONE
-        );
-
-        self::assertTrue(isset($object->blim));
-    }
-
-    /**
-     * See php isset documentation
-     *
-     * @test
-     * @return void
-     */
-    public function defined_to_null_property_returns_isset_false(): void
-    {
-        $object = new DataTransferObject(
-            ['blim' => $this->createMock(Property::class)],
-            ['blim' => null],
-            NONE
-        );
-
-        self::assertFalse(isset($object->blim));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function defined_to_anything_properties_return_is_defined_true(): void
-    {
-        $object = new DataTransferObject(
-            [
-                'blim' => $this->createMock(Property::class),
-                'blam' => $this->createMock(Property::class),
-            ],
-            [
-                'blim' => null,
-                'blam' => true
-            ],
-            NONE
-        );
-
-        self::assertTrue($object->isDefined('blim'));
-        self::assertTrue($object->isDefined('blam'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function is_defined_supports_dot_notation_for_nested_properties(): void
-    {
-        $object = new DataTransferObject(
-            ['blim' => $this->createMock(Property::class)],
-            ['blim' => new DataTransferObject(
-                ['blam' => $this->createMock(Property::class)],
-                ['blam' => new DataTransferObject(
-                    ['beep' => $this->createMock(Property::class)],
-                    ['beep' => true],
-                    NONE
-                )],
-                NONE
-            )],
-            NONE
-        );
-
-        self::assertTrue($object->isDefined('blim.blam.beep'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_property_returns_is_defined_false(): void
-    {
-        $object = new DataTransferObject(
-            ['blim' => $this->createMock(Property::class)],
-            [],
-            NONE
-        );
-
-        self::assertFalse($object->isDefined('blim'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function unknown_property_returns_is_defined_false(): void
-    {
-        $this->expectException(UnknownPropertiesError::class);
-
-        $object = new DataTransferObject(
-            ['blim' => $this->createMock(Property::class)],
-            [],
-            NONE
-        );
-
-        self::assertFalse($object->isDefined('blam'));
-    }
 
     /**
      * @test
