+++ tests/Unit/FactoryTest.php
@@ -4,24 +4,15 @@ declare(strict_types=1);
 
 namespace Rexlabs\DataTransferObject\Tests\Unit;
 
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Rexlabs\DataTransferObject\ClassData;
 use Rexlabs\DataTransferObject\DataTransferObject;
 use Rexlabs\DataTransferObject\DTOMetadata;
-use Rexlabs\DataTransferObject\Exceptions\ImmutableError;
-use Rexlabs\DataTransferObject\Exceptions\UninitialisedPropertiesError;
 use Rexlabs\DataTransferObject\Factory;
 use Rexlabs\DataTransferObject\Property;
 
 use function spl_object_id;
 
-use const Rexlabs\DataTransferObject\ARRAY_DEFAULT_TO_EMPTY_ARRAY;
-use const Rexlabs\DataTransferObject\BOOL_DEFAULT_TO_FALSE;
-use const Rexlabs\DataTransferObject\MUTABLE;
 use const Rexlabs\DataTransferObject\NONE;
-use const Rexlabs\DataTransferObject\NULLABLE_DEFAULT_TO_NULL;
-use const Rexlabs\DataTransferObject\PARTIAL;
 
 /**
  * Class FactoryTest
@@ -95,384 +86,5 @@ class FactoryTest extends TestCase
         );
 
         self::assertEquals($object->getDefinedProperties(), $properties);
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function creates_immutable_properties_by_default(): void
-    {
-        $this->expectException(ImmutableError::class);
-
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'string'], [], false, null)],
-            DataTransferObject::class,
-            ['one' => 'One'],
-            NONE
-        );
-
-        $object->__set('one', 'mutation');
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function creates_mutable_properties_when_specified(): void
-    {
-        $newValue = 'mutation';
-
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'string'], [], false, null)],
-            DataTransferObject::class,
-            ['one' => 'One'],
-            MUTABLE
-        );
-
-        $object->__set('one', $newValue);
-
-        self::assertEquals($newValue, $object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_nullable_property_returns_null(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'string'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            NULLABLE_DEFAULT_TO_NULL
-        );
-        $data = $object->toArray();
-
-        self::assertArrayHasKey('one', $data);
-        self::assertNull($data['one']);
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function can_make_partial_without_required_fields(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['string'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            PARTIAL
-        );
-
-        self::assertNotEmpty($object);
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_nullable_property_on_partial_returns_null(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['string'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            PARTIAL
-        );
-
-        self::assertNull($object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_nullable_property_omitted_by_to_array(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['string'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            PARTIAL
-        );
-
-        self::assertArrayNotHasKey('one', $object->toArray());
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function undefined_non_nullable_property_throws(): void
-    {
-        $this->expectException(UninitialisedPropertiesError::class);
-
-        $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['string'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            NULLABLE_DEFAULT_TO_NULL | ARRAY_DEFAULT_TO_EMPTY_ARRAY
-        );
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function array_defaults_to_empty_array(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['array'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            NULLABLE_DEFAULT_TO_NULL | ARRAY_DEFAULT_TO_EMPTY_ARRAY
-        );
-
-
-        self::assertEquals([], $object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function bool_defaults_to_false(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['bool'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            BOOL_DEFAULT_TO_FALSE
-        );
-
-
-        self::assertEquals(false, $object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function empty_array_takes_precedence_over_nullable(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'array'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            NULLABLE_DEFAULT_TO_NULL | ARRAY_DEFAULT_TO_EMPTY_ARRAY
-        );
-
-        self::assertEquals([], $object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function default_takes_precedence_over_nullable_or_empty_array(): void
-    {
-        $object = $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'array'], [], true, 'blim')],
-            DataTransferObject::class,
-            [],
-            NULLABLE_DEFAULT_TO_NULL | ARRAY_DEFAULT_TO_EMPTY_ARRAY
-        );
-
-
-        self::assertEquals('blim', $object->__get('one'));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function nullable_and_array_defaults_ignored_without_flags(): void
-    {
-        $this->expectException(UninitialisedPropertiesError::class);
-
-        $this->factory->makeWithProperties(
-            ['one' => new Property($this->factory, 'one', ['null', 'array', 'bool'], [], false, null)],
-            DataTransferObject::class,
-            [],
-            NONE
-        );
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function partial_flags_allows_missing_properties(): void
-    {
-        $data = ['one' => 1];
-        $object = $this->factory->makeWithProperties(
-            [
-                'one' => new Property($this->factory, 'one', ['int'], [], false, null),
-                'two' => new Property($this->factory, 'two', ['string'], [], false, null),
-            ],
-            DataTransferObject::class,
-            $data,
-            PARTIAL
-        );
-
-        self::assertEquals($data, $object->toArray());
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function extract_use_statements_with_aliases(): void
-    {
-        $code = <<< 'TEXT'
-<?php
-
-declare(strict_types=1);
-
-namespace Rexlabs\DataTransferObject\Tests\TransferObjects;
-
-use Acme\TestingPhoneDto as Phone;
-use Acme\TestingAddressDto;
-
-class TestingPersonDto extends DataTransferObject
-TEXT;
-
-        $expected = [
-            'Phone' => 'Acme\\TestingPhoneDto',
-            'TestingAddressDto' => 'Acme\\TestingAddressDto',
-        ];
-
-        self::assertEquals($expected, $this->factory->extractUseStatements($code));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function extract_use_statements_with_leading_slashes(): void
-    {
-        $code = <<< 'TEXT'
-<?php
-
-declare(strict_types=1);
-
-namespace Rexlabs\DataTransferObject\Tests\TransferObjects;
-
-use \Acme\TestingAddressDto;
-
-class TestingPersonDto extends DataTransferObject
-TEXT;
-
-        $expected = [
-            'TestingAddressDto' => 'Acme\\TestingAddressDto',
-        ];
-
-        self::assertEquals($expected, $this->factory->extractUseStatements($code));
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function extract_property_types_from_class_data(): void
-    {
-        $useStatements = [
-            'Phone' => 'Test\\TestingPhoneDto',
-            'TestingAddressDto' => 'Test\\TestingAddressDto',
-        ];
-
-        $docComment = <<<'TEXT'
-/**
- * @property string $first_name
- * @property null|string $last_name
- * @property string[] $aliases
- * @property-read null|Phone $phone
- * @property-read null|string $email
- * @property-read null|TestingAddressDto $address
- * @property-read null|Test\TestingAddressDto $postal_address
- * @property-read null|TestingAddressDto[] $other_addresses
- * @property-read string $status
- */
-TEXT;
-
-        $classData = new ClassData(
-            'Test\\Namespace\\DTO',
-            '',
-            $docComment,
-            [],
-            NONE
-        );
-
-        /**
-         * @var Property $firstName
-         * @var Property $lastName
-         * @var Property $aliases
-         * @var Property $phone
-         * @var Property $email
-         * @var Property $address
-         * @var Property $postalAddress
-         * @var Property $otherAddresses
-         * @var Property $status
-         */
-        [
-            $firstName,
-            $lastName,
-            $aliases,
-            $phone,
-            $email,
-            $address,
-            $postalAddress,
-            $otherAddresses,
-            $status,
-        ] = array_values($this->factory->mapClassToPropertyTypes($classData, $useStatements));
-
-        self::assertEquals(['string'], $firstName->getTypes());
-        self::assertEquals(['null', 'string'], $lastName->getTypes());
-        self::assertEquals(['string[]'], $aliases->getTypes());
-        self::assertEquals(['string'], $aliases->getArrayTypes());
-        self::assertEquals(['null', 'Test\\TestingPhoneDto'], $phone->getTypes());
-        self::assertEquals(['null', 'string'], $email->getTypes());
-        self::assertEquals(['null', 'Test\\TestingAddressDto'], $address->getTypes());
-        self::assertEquals(['null', 'Test\\TestingAddressDto'], $postalAddress->getTypes());
-        self::assertEquals(['null', 'Test\\TestingAddressDto[]'], $otherAddresses->getTypes());
-        self::assertEquals(['string'], $status->getTypes());
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function can_map_simple_types(): void
-    {
-        $type = $this->factory->mapType('Phone', null, ['Phone' => 'Acme\\Test\\TestingPhoneDto']);
-
-        self::assertEquals('Acme\\Test\\TestingPhoneDto', $type);
-    }
-
-    /**
-     * @test
-     * @return void
-     */
-    public function can_map_type_with_leading_slash(): void
-    {
-        /**
-         * @var Factory|MockObject $factory
-         */
-        $factory = $this->getMockBuilder(Factory::class)
-            ->setConstructorArgs([[]])
-            ->onlyMethods(['classExists'])
-            ->getMock();
-
-        $factory->method('classExists')->willReturn(true);
-
-        $type = $factory->mapType(
-            '\\Acme\\Test\\TestingPhoneDto',
-            null,
-            ['Phone' => 'Acme\\Test\\TestingPhoneDto']
-        );
-
-        self::assertEquals('Acme\\Test\\TestingPhoneDto', $type);
     }
 }
